Meta+LibWeb: Port PropertyID.h/cpp generators to invoke_generator()
