+++ Meta/Lagom/Tools/CodeGenerators/LibWeb/CMakeLists.txt
@@ -1,8 +1,7 @@
 set(SOURCES "") # avoid pulling SOURCES from parent scope
 
 lagom_tool(GenerateCSSMediaFeatureID       SOURCES GenerateCSSMediaFeatureID.cpp LIBS LagomMain)
-lagom_tool(Generate_CSS_PropertyID_h       SOURCES Generate_CSS_PropertyID_h.cpp LIBS LagomMain)
-lagom_tool(Generate_CSS_PropertyID_cpp     SOURCES Generate_CSS_PropertyID_cpp.cpp LIBS LagomMain)
+lagom_tool(GenerateCSSPropertyID           SOURCES GenerateCSSPropertyID.cpp LIBS LagomMain)
 lagom_tool(GenerateCSSValueID              SOURCES GenerateCSSValueID.cpp LIBS LagomMain)
 
 add_subdirectory(WrapperGenerator)
